Update: mychannel라우터, 컨트롤러 변경
